#00064 #工程文件提交

+++ Assets/Scripts/UI/HUD/HUDBuff.cs
@@ -32,7 +32,7 @@ namespace WarGame.UI
         //    }
         //}
 
-        public void UpdateBuff(int id, int leftDuration)
+        public void UpdateBuff(int id, float leftDuration)
         {
             var buffConfig = ConfigMgr.Instance.GetConfig<BufferConfig>("BufferConfig", id);
             _title.text = buffConfig.Name;
